Add security layer options and implementation
This commit adds command-line flags to the RepyV2 sandbox that allow
for donors and owners to set security layers, and also specifiy which
directories to read these security layers from:

--donor-sec-dir, --donor-sec-layers, --owner-sec-dir, --owner-sec-layers

(A bit of trickery is required so that the proper `chdir`s are issued.)

Implementation-wise, we mostly loop over the existing functionality in
`repy.py` that reads in user code, builds a `VirtualNamespace` out of it,
then evaluates it. In particular, we augment the function
`execute_namespace_until_completion` to return the evaluated namespace's
result context. This context is used as the base context (containing
possibly-overriden API definitions, but also additional functions, classes,
variables) for the next party's code.

Through this, the device donor defines what the vessel owner sees in
their `_context`, and likewise for the vessel owner and user.