Fix SeattleTestbed/repy_v2#98, GETTID on MIPS
This commit fixes an issue with the GETTID syscall on MIPS-based
systems such as most of our OpenWrt routers.

At the same time, it *creates* an issue on non-x86 and non-MIPS
systems, as I don't know and can't test the GETTID for, say,
ARMs right now. A friendly TODO comment points out this situation.