Restore `linux_api`'s GETTID for ARM targets
This is a follow-up commit to 643b912d9b02d160477086d03adc45195f44b7f8
which fixed SeattleTestbed/repy_v2#98 (GETTID on MIPS) but broke it
for ARMs at the same time. GETTID's syscall number on ARM is identical
with x86_32's, or so my testing on a Raspberry Pi Model B Rev 2 with
BCM2835 SoC with an ARM CPU core indicates.

This commit also adds raising an exception to signal that we have no
GETTID number in case we are running on a platform other than
x86_32/64, MIPS, or ARM.