Fix #126, namespace traversal for dict values
As documented by SeattleTestbed/repy_v2#126, namespace wrapping and
unwrapping for `dict`s is incomplete and breaks if a `dict` that
contains a value of certain types like `safe_type` or `classobj`
traverses the namespace in a RepyV2 API call. (At the moment, the
`evaluate(context)` method of VirtualNamespace is the only API
call that takes a `dict` as the parameter.)

This commit patches the relevant `_copy` method in `namespace.py`
to enable proper copying of said types. Furthermore, it adds a
unit test that checks many "easy", natural and often-encountered
types, but also checks "unusual" cases (like those above) to
verify that the patch really fixes the problem.